refactor: rename src/export/ to src/exporter/
Rename the export module directory to exporter for consistency with the
project name (bus-exporter) and crate conventions. Updates the module
declaration and all references in src/main.rs.

Pure rename + import update — no logic changes.

Closes #84
